server/requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
gcloud==0.18.3
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jws==0.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.1
python-jwt==2.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
urllib3==1.26.18
Werkzeug==3.0.1
wheel==0.42.0
